remove binder
