+++ mobile/lib/main.dart
@@ -1,8 +1,6 @@
-import 'package:binder/binder.dart';
 import 'package:flutter/material.dart';
 import 'package:mobile/routes/home_route.dart';
 import 'package:mobile/state.dart';
-import 'package:mobile/state.momentum.dart';
 import 'package:momentum/momentum.dart';
 
 void main() {
@@ -16,19 +14,12 @@ class MyApp extends StatelessWidget {
   Widget build(BuildContext context) {
     return Momentum(
       controllers: [AppController()],
-      child: BinderScope(
-        // TODO: log out state transitions
-        // observers: [DelegatingStateObserver((ref){})],
-        child: LogicLoader(
-          refs: [appViewLogicRef],
-          child: MaterialApp(
-            title: 'Flutter Demo',
-            theme: ThemeData(
-              primarySwatch: Colors.blue,
-            ),
-            home: const HomeRoute(),
-          ),
+      child: MaterialApp(
+        title: 'Flutter Demo',
+        theme: ThemeData(
+          primarySwatch: Colors.blue,
         ),
+        home: const HomeRoute(),
       ),
     );
   }
