+++ mobile/pubspec.yaml
@@ -10,7 +10,6 @@ dependencies:
   flutter:
     sdk: flutter
 
-  binder: any
   cupertino_icons: ^1.0.2
   figma_squircle: any
   http: any
